chore(webpack-scaffold): : moved ts files to src folder and compile to lib
