+++ packages/webpack-scaffold/.gitignore
@@ -1,3 +1,4 @@
 *.js
 !*.test.js
 !/**/*.test.js
+lib/
