+++ packages/webpack-scaffold/package.json
@@ -2,8 +2,8 @@
   "name": "@webpack-cli/webpack-scaffold",
   "version": "1.0.1-alpha.0",
   "description": "Utility files for webpack-scaffold",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "main": "lib/index.js",
+  "types": "lib.index.d.ts",
   "publishConfig": {
     "access": "public"
   },
@@ -20,5 +20,8 @@
     "test": "jest --detectOpenHandles",
     "build": "tsc",
     "watch": "npm run build && tsc -w"
-  }
+  },
+  "files": [
+    "lib"
+  ]
 }
