+++ packages/webpack-scaffold/src/index.ts
